requirements.txt
# Core dependencies
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
yt-dlp==2024.1.1
pydub==0.25.1
ffmpeg-python==0.2.1
Pillow==10.1.0
numpy==1.24.3

# AI/ML
google-generativeai==0.3.1

# Data & Database
pandas==2.1.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
